PyHwpServer/pyproject.toml
[tool.poetry]
name = "pyhwpserver"
version = "0.1.0"
description = ""
authors = ["Flzomb <spark3840@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pywin32 = "^306"
pyyaml = "^6.0.2"
fastapi = {extras = ["standard"], version = "^0.115.0"}
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
